+++ Runtime/MicroActor/MicroActorSpawner.cs
@@ -8,11 +8,12 @@ namespace Pixygon.Micro {
 
         private GameObject _spawnedActor;
         
-        public async void SpawnActor() {
+        public async void SpawnActor(LevelLoader loader) {
             if(_spawnedActor != null) DespawnActor();
             Log.DebugMessage(DebugGroup.Actor, "Spawning actor", this);
             _spawnedActor = await AddressableLoader.LoadGameObject(_actorData._actorRef, transform);
             _spawnedActor.transform.localPosition = Vector3.zero;
+            _spawnedActor.GetComponent<MicroActor>().Initialize(loader, _actorData);
         }
 
         private void DespawnActor() {