+++ Runtime/DamageObject.cs
@@ -1,0 +1,5 @@
+﻿using UnityEngine;
+
+namespace Pixygon.Micro {
+    public class DamageObject : MonoBehaviour { }
+}