+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## Unreleased
+### Fixed
+- Run-particles should be off on death now
+- Run-particles should only appear when actually moving now
+
 ## 0.5.4
 ### Fixed
 - MicroActorSpawners spawns enemies inside it